setup REST api
