+++ package.json
@@ -4,7 +4,8 @@
   "description": "\"# task-app\"",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node src/index.js",
+    "dev": "nodemon src/index.js"
   },
   "repository": {
     "type": "git",
@@ -17,6 +18,7 @@
   },
   "homepage": "https://github.com/gitdeepsingh/task-app#readme",
   "dependencies": {
+    "express": "^4.17.1",
     "mongodb": "^3.3.2",
     "mongoose": "^5.7.7",
     "validator": "^12.0.0"
