+++ src/db/mongoose.js
@@ -1,82 +1,7 @@
 const mongoose = require('mongoose');
-const validator = require('validator');
 
 mongoose.connect('mongodb://127.0.0.1:27017/task-manager-api', {
     useNewUrlParser: true,
     useUnifiedTopology: true
     // useCreateIndex: true
-});
-
-const User = mongoose.model('User', {
-    name: {
-        type: String,
-        required: true,
-        trim: true
-    },
-    email: {
-        type: String,
-        required: true,
-        trim: true,
-        lowercase: true,
-        validate(value) {
-            if (!validator.isEmail(value)) {
-                throw new Error('Email is invalid!');
-            }
-        }
-    },
-    password: {
-        type: String,
-        required: true,
-        minlength: 7,
-        trim: true,
-        validate(value) {
-            if (value.toLowerCase().includes('password')) {
-                throw new Error('Password cannot contain "password"')
-            }
-        }
-    },
-    age: {
-        type: Number,
-        default: 0,
-        validate(value) {
-            if (value < 0) throw new Error("Age cannnot be negative!")
-        }
-    }
-});
-
-const Task = mongoose.model('Task', {
-    description: {
-        type: String,
-        required: true,
-        trim: true
-    },
-    completed: {
-        type: Boolean,
-        default: false
-    }
-})
-
-const me = new User({
-    name: 'Deep Singh',
-    email: 'deepsingh@gmail.com',
-    password: 'Bang123',
-    age: 27
-});
-
-const myTask = new Task({
-    description: 'Prepare Google sheet for sync logs report.',
-    completed: false
-});
-
-
-// me.save().then(() => {
-//     console.log('me==', me)
-// }).catch((err) => {
-//     console.log('Error: ', err)
-// });
-
-myTask.save().then(() => {
-    console.log('task==', myTask)
-}).catch((err) => {
-    console.log('Error: ', err)
 });
